+++ WireDomain/Sources/WireDomainPackage/UseCases/Protocols/CreateLegacyBackupError.swift
