+++ WireDomain/Sources/WireDomainPackageSupport/Sourcery/AutoMockable.stencil
