+++ WireDomain/Sources/WireDomainPackageSupport/Sourcery/sourcery.yml
@@ -1,0 +1,8 @@
+sources:
+- ${PACKAGE_ROOT_DIR}/Sources/WireDomainPackage
+templates:
+- ${TARGET_DIR}/Sourcery/AutoMockable.stencil
+output:
+  ${DERIVED_SOURCES_DIR}
+args:
+  autoMockablePublicImports: ["Foundation", "WireDomainPackage"]
