--- WireDomain/Sources/WireDomainPkgSupport/Sourcery/sourcery.yml
@@ -1,8 +1,0 @@
-sources:
-- ${PACKAGE_ROOT_DIR}/Sources/WireDomainPkg
-templates:
-- ${TARGET_DIR}/Sourcery/AutoMockable.stencil
-output:
-  ${DERIVED_SOURCES_DIR}
-args:
-  autoMockablePublicImports: ["Foundation", "WireDomainPkg"]
