+++ WireDomain/Sources/WireDomainSupport/Sourcery/config.yml
@@ -5,5 +5,5 @@ templates:
 output:
   ./generated
 args:
-  autoMockableImports: ["WireAPI", "WireDataModel", "WireDomainPkg"]
+  autoMockableImports: ["WireAPI", "WireDataModel", "WireDomainPackage"]
   autoMockableTestableImports: ["WireDomain"]
