+++ WireDomain/Tests/TestPlans/AllTests.xctestplan
@@ -33,8 +33,8 @@
     {
       "target" : {
         "containerPath" : "container:",
-        "identifier" : "WireDomainPkgTests",
-        "name" : "WireDomainPkgTests"
+        "identifier" : "WireDomainPackageTests",
+        "name" : "WireDomainPackageTests"
       }
     },
     {
