+++ WireUI/Sources/WireIndividualToTeamMigrationUI/Views/PreviewsSetup/_FeaturePreview.swift
@@ -17,7 +17,7 @@
 //
 
 import SwiftUI
-import WireDomainPkg
+import WireDomainPackage
 
 public struct MockUseCase: IndividualToTeamMigrationUseCaseProtocol {
     let error: (any Error)?
