+++ WireUI/Sources/WireSettingsUI/Account/BackupImportExport/Import/ImportBackupViewModel.swift
@@ -17,7 +17,7 @@
 //
 
 import Foundation
-import WireDomainPkg
+import WireDomainPackage
 import WireLogging
 
 @MainActor
