+++ WireUI/Sources/WireSettingsUI/Account/BackupImportExport/Previews/PreviewImportBackupUseCase.swift
@@ -17,7 +17,7 @@
 //
 
 import Foundation
-import WireDomainPkg
+import WireDomainPackage
 
 struct PreviewImportBackupUseCase: ImportBackupUseCaseProtocol {
