+++ wire-ios-sync-engine/Source/Use cases/ImportBackupUseCase/ImportBackupFileArchiver.swift
@@ -16,7 +16,7 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import WireDomainPkg
+import WireDomainPackage
 import ZipArchive
 
 struct ImportBackupFileArchiver: ImportBackupFileArchiverProtocol {
