+++ wire-ios-sync-engine/Tests/Source/Use cases/ImportBackupUseCaseTests.swift
@@ -17,7 +17,7 @@
 //
 
 import WireDataModelSupport
-import WireDomainPkg
+import WireDomainPackage
 import XCTest
 
 @testable import WireSyncEngine
