+++ wire-ios/Wire-iOS/Sources/UserInterface/Settings/Backup/CreateLegacyBackupUseCase.swift
@@ -17,7 +17,7 @@
 //
 
 import Foundation
-import WireDomainPkg
+import WireDomainPackage
 import WireSettingsUI
 import WireSyncEngine
